Use PerScopeLifetime rather than PerRequestLifetime
Quote: "The PerRequestLifetime gives you a new instance per request where request represents calling GetInstance on the container. The lifetime you are looking for is the PerScopeLifetime where one instance is produced for each scope. The Scope starts when the web request starts and ends when the web request ends." https://github.com/seesharper/LightInject/issues/131